+++ src/AdapterBackblazeS3.ts
@@ -1,20 +1,21 @@
 import { S3Client, _Object, GetBucketAclCommand, PutObjectCommand, DeleteObjectCommand, } from "@aws-sdk/client-s3";
 import { Options, Provider } from "./types/general";
 import { ResultObject, ResultObjectBoolean, ResultObjectObject, } from "./types/result";
-import { AdapterConfigAmazonS3 } from "./types/adapter_amazon_s3";
+import { AdapterConfigS3 } from "./types/adapter_amazon_s3";
 import { getErrorMessage } from "./util";
 import { AdapterAmazonS3 } from "./AdapterAmazonS3";
 import { getSignedUrl } from "@aws-sdk/s3-request-presigner";
 
 export class AdapterBackblazeS3 extends AdapterAmazonS3 {
-  declare protected _config: AdapterConfigAmazonS3;
+  declare protected _config: AdapterConfigS3;
   declare protected _client: S3Client;
   protected _provider: Provider = Provider.BACKBLAZE_S3;
   protected _configError: null | string = null;
 
-  constructor(config: string | AdapterConfigAmazonS3) {
+  constructor(config: string | AdapterConfigS3) {
     super(config);
     this.parseConfig(config)
+    this.checkConfig();
     this.createClient();
   }
 
@@ -101,5 +102,13 @@ export class AdapterBackblazeS3 extends AdapterAmazonS3 {
     } catch (e: unknown) {
       return { value: null, error: getErrorMessage(e) };
     }
+  }
+
+  get config(): AdapterConfigS3 {
+    return this._config as AdapterConfigS3;
+  }
+
+  getConfig(): AdapterConfigS3 {
+    return this._config as AdapterConfigS3;
   }
 }