+++ src/types/adapter_amazon_s3.ts
@@ -10,3 +10,14 @@ export interface AdapterConfigAmazonS3 extends AdapterConfig {
   accessKeyId?: string;
   secretAccessKey?: string;
 }
+
+export interface AdapterConfigS3 extends AdapterConfigAmazonS3 {
+  region?: string;
+  endpoint: string;
+  credentials?: {
+    accessKeyId?: string;
+    secretAccessKey?: string;
+  };
+  accessKeyId: string;
+  secretAccessKey: string;
+}