+++ tests/config.ts
@@ -39,7 +39,7 @@ export function getConfig(provider: string = Provider.LOCAL): string | StorageAd
       accessKeyId: process.env.R2_ACCESS_KEY_ID,
       secretAccessKey: process.env.R2_SECRET_ACCESS_KEY,
     };
-    config = `r2://${process.env.R2_ACCESS_KEY_ID}:${process.env.R2_SECRET_ACCESS_KEY}@${process.env.BUCKET_NAME}?endpoint=${process.env.R2_ENDPOINT}&region=${process.env.R2_REGION}`;
+    // config = `r2://${process.env.R2_ACCESS_KEY_ID}:${process.env.R2_SECRET_ACCESS_KEY}@${process.env.BUCKET_NAME}?endpoint=${process.env.R2_ENDPOINT}&region=${process.env.R2_REGION}`;
   } else if (provider === Provider.B2_S3) {
     config = {
       provider,
@@ -52,7 +52,7 @@ export function getConfig(provider: string = Provider.LOCAL): string | StorageAd
   } else if (provider === Provider.CUBBIT) {
     config = {
       provider,
-      // region: "eu",
+      region: "us-east-1",
       bucketName: process.env.BUCKET_NAME,
       endpoint: process.env.CUBBIT_ENDPOINT,
       accessKeyId: process.env.CUBBIT_ACCESS_KEY_ID,
@@ -67,6 +67,7 @@ export function getConfig(provider: string = Provider.LOCAL): string | StorageAd
       // accessKeyId: process.env.MINIO_ACCESS_KEY_DOCKER,
       // secretAccessKey: process.env.MINIO_SECRET_KEY_DOCKER,
 
+      region: "us-east-1",
       endpoint: process.env.MINIO_ENDPOINT_S3,
       accessKeyId: process.env.MINIO_ACCESS_KEY,
       secretAccessKey: process.env.MINIO_SECRET_KEY,