+++ publish/AdapterAmazonS3/changelog.md
@@ -1,11 +1,12 @@
-# 2.0.0
+# 3.0.0
+- version bump to match the version of Storage and the API
 - removed support for versioning (wasn't implemented consistently anyway)
 - removed `getFileAsURL`
 - removed option `{allVersions: boolean}` for `removeFile`
 - `removeFile` does not fail if the file doesn't exist
 - `deleteBucket` does not fail if the bucket doesn't exist
 - see also the [migration guide](https://github.com/tweedegolf/storage-abstraction/blob/master/migration_to_api3.0.md)
-- update @aws-sdk to 3.892.0
+- update @aws-sdk to 3.913.0
 
 # 1.0.18
 - update @aws-sdk to 3.888.0